+++ example/pubspec.yaml
@@ -25,8 +25,6 @@ dependencies:
     sdk: flutter
   image_painter:
     path: ../
-  flutter_material_color_picker: ^1.0.5
-  image_picker: ^0.6.7+14
   open_file: ^3.0.3
   path_provider: ^1.6.24
