+++ lib/image_painter.dart
@@ -1,3 +1,3 @@
 library image_painter;
 
-export 'src/paint_over_image.dart';
+export 'src/_paint_over_image.dart';
